+++ packages/react/src/components/button.tsx
@@ -1,30 +1,30 @@
 "use client";
 
+import { opacify } from "polished";
 import styled, { css } from "styled-components";
-import { theme } from "../styles";
 
 const baseStyles = css`
 	all: unset;
-	border-radius: ${props => theme.radii.sm};
-	font-size: ${props => theme.fontSizes.sm};
-	font-weight: ${props => theme.fontWeights.medium};
-	font-family: ${props => theme.fonts.default};
+	border-radius: ${props => props.theme.radii.sm};
+	font-size: ${props => props.theme.fontSizes.sm};
+	font-weight: ${props => props.theme.fontWeights.medium};
+	font-family: ${props => props.theme.fonts.default};
 	text-align: center;
 	min-width: 120px;
-	padding: 0 ${props => theme.space[4]};
+	padding: 0 ${props => props.theme.space[4]};
 	box-sizing: border-box;
 
 	display: flex;
 	align-items: center;
 	justify-content: center;
-	gap: ${props => theme.space[2]};
+	gap: ${props => props.theme.space[2]};
 
 	cursor: pointer;
-	transition: ${props => theme.transitions[100]};
+	transition: ${props => props.theme.transitions[100]};
 
 	svg {
-		width: ${props => theme.space[4]};
-		height: ${props => theme.space[4]};
+		width: ${props => props.theme.space[4]};
+		height: ${props => props.theme.space[4]};
 	}
 
 	&:not(:disabled):active {
@@ -36,98 +36,96 @@ const baseStyles = css`
 	}
 
 	&:focus-visible {
-		box-shadow: 0 0 0 2px ${props => theme.colors.gray900};
+		box-shadow: 0 0 0 2px ${props => props.theme.colors.gray900};
 	}
 `;
 
 const variantStyles = {
 	primary: css`
-		color: ${props => theme.colors.white300};
-		background-color: ${props => theme.colors.unimake300};
+		color: ${props => props.theme.colors.white300};
+		background-color: ${props => props.theme.colors.unimake300};
 
 		&:not(:disabled):hover {
-			background-color: ${props => theme.colors.unimake500};
+			background-color: ${props => props.theme.colors.unimake500};
 		}
 
 		&:disabled {
-			background-color: ${props => theme.colors.gray200};
+			opacity: 0.6;
 		}
 	`,
 	secondary: css`
-		color: ${props => theme.colors.gray900};
-		background-color: ${props => theme.colors.gray100};
+		color: ${props => props.theme.colors.gray900};
+		background-color: ${props => props.theme.colors.gray100};
 
 		&:not(:disabled):hover {
-			background-color: ${props => theme.colors.gray200};
+			background-color: ${props => props.theme.colors.gray200};
 		}
 
 		&:disabled {
-			opacity: 0.5;
+			opacity: 0.6;
 		}
 	`,
 
 	tertiary: css`
-		color: ${props => theme.colors.gray600};
+		color: ${props => props.theme.colors.gray500};
 
 		&:not(:disabled):hover {
-			color: ${props => theme.colors.black};
+			color: ${props => props.theme.colors.gray900};
 		}
 
 		&:disabled {
-			color: ${props => theme.colors.gray400};
+			opacity: 0.6;
 		}
 	`,
 
 	danger: css`
-		color: ${props => theme.colors.white300};
-		background-color: ${props => theme.colors.red500};
+		color: ${props => props.theme.colors.white300};
+		background-color: ${props => props.theme.colors.red500};
 
 		&:not(:disabled):hover {
-			background-color: ${props => theme.colors.red700};
+			background-color: ${props => props.theme.colors.red700};
 		}
 
 		&:disabled {
-			background-color: ${props => theme.colors.gray200};
+			opacity: 0.6;
 		}
 	`,
 
 	safe: css`
-		color: ${props => theme.colors.white300};
-		background-color: ${props => theme.colors.green500};
+		color: ${props => props.theme.colors.white300};
+		background-color: ${props => props.theme.colors.green500};
 
 		&:not(:disabled):hover {
-			background-color: ${props => theme.colors.green700};
+			background-color: ${props => props.theme.colors.green700};
 		}
 
 		&:disabled {
-			background-color: ${props => theme.colors.gray200};
+			opacity: 0.6;
 		}
 	`,
 
 	ghost: css`
-		color: ${props => theme.colors.gray900};
+		color: ${props => props.theme.colors.gray900};
 
 		&:not(:disabled):hover {
-			background-color: ${props => theme.colors.gray100};
+			background-color: ${props => opacify(-0.5, props.theme.colors.gray100)};
 		}
 
 		&:disabled {
-			color: ${props => theme.colors.gray600};
-			opacity: 0.5;
+			opacity: 0.6;
 		}
 	`,
 
 	link: css`
 		min-width: unset;
-		color: ${props => theme.colors.gray900};
+		color: ${props => props.theme.colors.gray900};
 
 		&:not(:disabled):hover {
 			text-decoration: underline;
 		}
 
 		&:disabled {
-			color: ${props => theme.colors.gray600};
-			opacity: 0.5;
+			opacity: 0.6;
 		}
 
 		&:focus-visible {
@@ -154,15 +152,15 @@ const sizeStyles = {
 };
 
 export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-	variant?: "primary" | "secondary" | "tertiary" | "danger" | "safe" | "ghost" | "link";
-	size?: "sm" | "md" | "full";
+	$variant?: "primary" | "secondary" | "tertiary" | "danger" | "safe" | "ghost" | "link";
+	$size?: "sm" | "md" | "full";
 }
 
 export const Button = styled.button<ButtonProps>`
 	${baseStyles}
-	${({ variant = "primary", size = "md" }) => css`
-		${variantStyles[variant]}
-		${sizeStyles[size]}
+	${({ $variant = "primary", $size = "md" }) => css`
+		${variantStyles[$variant]}
+		${sizeStyles[$size]}
 	`}
 `;