+++ packages/react/src/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 export * from "./components/button";
 export * from "./components/heading";
 export * from "./components/text";