+++ packages/tokens/src/colors.ts
@@ -28,3 +28,34 @@ export const colors = {
 	red500: "#BF2234",
 	red700: "#9C1C2B"
 } as const;
+
+export const darkColors = {
+	black: "#000",
+
+	white100: "#FFF",
+	white300: "#FAFAFA",
+
+	gray100: "#121214",
+	gray200: "#202024",
+	gray300: "#29292E",
+	gray400: "#323238",
+	gray500: "#ACAFB8",
+	gray600: "#7C7C8A",
+	gray700: "#93939f",
+	gray800: "#C9C9D2",
+	gray900: "#E1E1E6",
+
+	unimake100: "#00527A",
+	unimake300: "#004465",
+	unimake500: "#003652",
+	unimake700: "#00293d",
+	unimake900: "#001b29",
+
+	green300: "#28A847",
+	green500: "#1F8438",
+	green700: "#17632A",
+
+	red300: "#DB3247",
+	red500: "#BF2234",
+	red700: "#9C1C2B"
+} as const;