+++ packages/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @unimake-ui/react
 
+## 1.0.4
+
+### Patch Changes
+
+- Export ThemeProvider and add darkColors to theme object; Uses props.theme inside all components; Add the prefix "$" in all component variables
+
+## 1.0.3
+
+### Patch Changes
+
+- Add 'use client' at index.ts file which is exporting all styled components
+
+## 1.0.2
+
+### Patch Changes
+
+- Add text and heading components and uses "use client" directive
+
 ## 1.0.1
 
 ### Patch Changes